Restore types of FastifyRequest - Move `/typings/fastify/index.d.ts` code to `/src/index.ts`. - Add some changes for satisfy the linter - Add Webstorm IDE's configuration files to .gitignore
