+++ package.json
@@ -3,7 +3,7 @@
   "description": "Fastify plugin for handling `multipart/form-data`.",
   "version": "2.0.2",
   "main": "lib/index.js",
-  "types": "lib/",
+  "types": "lib/index.d.ts",
   "contributors": [
     "Maksim Sinik <maksim@sinik.it>",
     "Hage Yaapa <captain@hacksparrow.com> (http://www.hacksparrow.com)",
